Rename color-rgb -> ColorRGB
