+++ packages/delir-core/src/Values/index.ts
@@ -2,6 +2,6 @@ export {default as Point2D} from './point-2d'
 export {default as Point3D} from './point-3d'
 export {default as Size2D} from './size-2d'
 export {default as Size3D} from './size-3d'
-export {default as ColorRGB} from './color-rgb'
+export {default as ColorRGB} from './ColorRGB'
 export {default as ColorRGBA} from './ColorRGBA'
 export {default as Expression} from './Expression'
